Move image and texture loaders into new 'loaders' directory
