+++ packages/graphics/src/loaders/ImageLoader.cpp
@@ -1,0 +1,2 @@
+#define STB_IMAGE_IMPLEMENTATION
+#include <graphics/loaders/ImageLoader.h>
