--- packages/graphics/src/textures/ImageLoader.cpp
@@ -1,2 +1,0 @@
-#define STB_IMAGE_IMPLEMENTATION
-#include <graphics/textures/ImageLoader.h>
